+++ src/IllyriaPlus.kt
@@ -7,10 +7,7 @@ import org.xodium.illyriaplus.enchantments.*
 import org.xodium.illyriaplus.mechanics.entity.*
 import org.xodium.illyriaplus.mechanics.player.*
 import org.xodium.illyriaplus.mechanics.server.*
-import org.xodium.illyriaplus.mechanics.world.ChiseledBookshelfMechanic
-import org.xodium.illyriaplus.mechanics.world.DimensionMechanic
-import org.xodium.illyriaplus.mechanics.world.InventoryMechanic
-import org.xodium.illyriaplus.mechanics.world.OpenableMechanic
+import org.xodium.illyriaplus.mechanics.world.*
 import org.xodium.illyriaplus.recipes.*
 
 /** Main class of the plugin. */
@@ -65,6 +62,7 @@ internal class IllyriaPlus : JavaPlugin() {
                 SpellMechanic,
                 TabListMechanic,
                 TameableMechanic,
+                TeleportMechanic,
                 XPMechanic,
             )
