+++ src/mechanics/player/ExperienceMechanic.kt
@@ -1,0 +1,45 @@
+@file:Suppress("ktlint:standard:no-wildcard-imports")
+
+package org.xodium.illyriaplus.mechanics.player
+
+import org.bukkit.Material
+import org.bukkit.event.EventHandler
+import org.bukkit.event.player.PlayerInteractEvent
+import org.bukkit.inventory.ItemStack
+import org.xodium.illyriaplus.interfaces.MechanicInterface
+
+/** Allows players to convert XP into experience bottles at enchanting tables. */
+internal object ExperienceMechanic : MechanicInterface {
+    const val XP_COST_TO_BOTTLE: Int = 11
+
+    @EventHandler
+    fun on(event: PlayerInteractEvent) {
+        xpToBottle(event)
+    }
+
+    /**
+     * Handles the interaction event where a player can convert their experience points into an experience bottle
+     * if specific conditions are met.
+     *
+     * @param event The PlayerInteractEvent triggered when a player interacts with the world or an object.
+     */
+    private fun xpToBottle(event: PlayerInteractEvent) {
+        if (event.clickedBlock?.type != Material.ENCHANTING_TABLE ||
+            event.item?.type != Material.GLASS_BOTTLE ||
+            !event.player.isSneaking
+        ) {
+            return
+        }
+
+        val player = event.player
+
+        if (player.calculateTotalExperiencePoints() < XP_COST_TO_BOTTLE) return
+
+        player.giveExp(-XP_COST_TO_BOTTLE)
+        event.item?.subtract(1)
+        player.inventory
+            .addItem(ItemStack.of(Material.EXPERIENCE_BOTTLE, 1))
+            .values
+            .forEach { player.world.dropItemNaturally(player.location, it) }
+    }
+}
